Add info field support to amadeusMenu
Selecting a configuration option now shows more details and asks whether user wants to change it or set it to its default if there is one. Previously it would ask for input immediately. Changed to allow for default option and to be more user friendly. Setting to default not implemented yet.